+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adds support for 204 no content in generation and CSharp/Java/Go/TypeScript request adapters. #1410
+- Adds text output formatter to CLI commons.
 
 ### Changed